+++ climpred/bias_removal.py
@@ -347,6 +347,7 @@ def gaussian_bias_removal(
             dim="member",
             alignment=alignment,
         )
+    bias = bias.drop("skill")
 
     # how to remove bias
     if "mean" in how: