+++ climpred/tests/test_HindcastEnsemble_class.py
@@ -2,8 +2,9 @@ import pytest
 import xarray as xr
 
 from climpred import HindcastEnsemble, set_options
-from climpred.exceptions import DimensionError
+from climpred.exceptions import CoordinateError, DimensionError
 from climpred.options import OPTIONS
+from climpred.utils import convert_time_index
 
 
 def test_hindcastEnsemble_init(hind_ds_initialized_1d):
@@ -274,3 +275,82 @@ def test_HindcastEnsemble_instantiating_standard_name(
     with pytest.warns(UserWarning, match="but renamed dimension"):
         init = HindcastEnsemble(init).get_initialized()
         assert dim in init.dims, print(init.dims, init.coords)
+
+
+@pytest.mark.parametrize(
+    "init_freq,lead_unit",
+    [
+        ("AS-JUL", "years"),
+        ("AS-JUL", "months"),
+        ("AS-JUL", "seasons"),
+        ("MS", "months"),
+        ("3M", "days"),
+        ("7D", "days"),
+        ("1D", "hours"),
+        ("1H", "seconds"),
+    ],
+)
+@pytest.mark.parametrize("calendar", ["ProlepticGregorian", "standard", "360_day"])
+def test_hindcastEnsemble_init_time(init_freq, lead_unit, calendar):
+    """Test to see HindcastEnsemble can be initialized and creates time
+    coordinate depending on init and lead for different calendars and lead units."""
+    p = 3
+    nlead = 2
+    lead = [0, 1]
+    import numpy as np
+
+    init = xr.cftime_range(start="2000", freq=init_freq, periods=p)
+    data = np.random.rand(p, nlead)
+    init = xr.DataArray(
+        data,
+        dims=["init", "lead"],
+        coords={"init": init, "lead": lead},
+        name="initialized",
+    )
+    init.lead.attrs["units"] = lead_unit
+    coords = HindcastEnsemble(init).coords
+    assert "valid_time" in coords
+    assert (coords["valid_time"].isel(lead=0, drop=True) == coords["init"]).all()
+    assert (coords["valid_time"].isel(lead=1, drop=True) != coords["init"]).all()
+
+
+@pytest.mark.parametrize("lead_freq", ["years", "months", "seasons"])
+def test_fractional_leads_360_day(hind_ds_initialized_1d, lead_freq):
+    """Test that lead can also contain floats when calendar='360_day'."""
+    hind_ds_initialized_1d["init"] = xr.cftime_range(
+        start=str(hind_ds_initialized_1d.init[0].values),
+        freq="YS",
+        periods=hind_ds_initialized_1d.init.size,
+        calendar="360_day",
+    )
+    with xr.set_options(keep_attrs=True):
+        hind_ds_initialized_1d["lead"] = hind_ds_initialized_1d["lead"] - 0.5
+        hind_ds_initialized_1d["lead"].attrs["units"] = lead_freq
+    assert HindcastEnsemble(hind_ds_initialized_1d)
+
+
+@pytest.mark.parametrize("lead_freq", ["weeks", "pentads", "days", "seconds"])
+@pytest.mark.parametrize(
+    "calendar", ["standard", "julian", "noleap", "proleptic_gregorian"]
+)
+def test_fractional_leads_lower_than_month_lead_units(
+    hind_ds_initialized_1d, lead_freq, calendar
+):
+    """Test that lead can also contain floats when lead units is lower or equal to weeks'."""
+    hind_ds_initialized_1d["init"] = xr.cftime_range(
+        start=str(hind_ds_initialized_1d.init[0].values),
+        freq="YS",
+        periods=hind_ds_initialized_1d.init.size,
+        calendar=calendar,
+    )
+    hind_ds_initialized_1d["lead"] = hind_ds_initialized_1d["lead"] - 0.5
+    hind_ds_initialized_1d["lead"].attrs["units"] = lead_freq
+    assert HindcastEnsemble(hind_ds_initialized_1d)
+
+
+def test_fractional_leads_fails(hind_ds_initialized_1d):
+    """Test that fractional leads fail for normal calendars and lead units in larger than days."""
+    with xr.set_options(keep_attrs=True):
+        hind_ds_initialized_1d["lead"] = hind_ds_initialized_1d["lead"] - 0.5
+    with pytest.raises(CoordinateError, match="Require integer"):
+        HindcastEnsemble(hind_ds_initialized_1d)