+++ climpred/tests/test_smoothing.py
@@ -173,6 +173,7 @@ def test_HindcastEnsemble_spatial_smoothing_dim_and_skill(hindcast_recon_3d, dim
     he = hindcast_recon_3d
     smooth_kws = {key: step for key in dim}
     he_smoothed = he.smooth(smooth_kws)
+    assert he_smoothed.get_initialized().lead.attrs is not None
     for d in dim:
         assert he_smoothed.get_initialized()[d].any()
         assert he_smoothed.get_observations()[d].any()